refactor(core): use QStandardPaths for tool detection and tighten types
Replace 'which' subprocess calls with QStandardPaths::findExecutable for
grim, oxipng/optipng and cwebp, use QImageWriter with explicit format and
quality in MetadataSanitizer, and adopt qsizetype/static_cast to silence
signedness warnings.

+++ src/core/LibraryManager.cpp
@@ -915,7 +915,7 @@ QVariantMap LibraryManager::findDuplicatesInLibrary(int tolerance) {
   res["totalReclaimableFormatted"] = formatFileSize(totalReclaim);
   res["groups"] = groupList;
 
-  emit duplicateScanFinished(groups.size(), totalReclaim);
+  emit duplicateScanFinished(static_cast<int>(groups.size()), totalReclaim);
   return res;
 }
